feat #25: 신고 클러스터가 제대로 합쳐지지 않는 문제 등 해결
- 정밀한 거리 계산을 위해 위경도의 타입을 Float에서 Double로 수정
- 거리 계산을 유클리드에서 하버사인으로 변경
- console log 출력 설정
- cluster count의 초기화 값을 1->0으로 변경 (실질적으로 저장되는 값이 1부터 시작하도록 함)

+++ backend/byeoltago-api/src/main/java/com/kakaoimpact/byeoltago_api/controller/ReportController.java
@@ -32,10 +32,10 @@ public class ReportController {
 
     // 위치 기반 마커 조회 API
     @GetMapping("/markers")
-    public ResponseEntity<List<ReportInfoResponseDto>> getMarkers(@RequestParam float minLat,
-                                                                 @RequestParam float maxLat,
-                                                                 @RequestParam float minLon,
-                                                                 @RequestParam float maxLon) {
+    public ResponseEntity<List<ReportInfoResponseDto>> getMarkers(@RequestParam double minLat,
+                                                                 @RequestParam double maxLat,
+                                                                 @RequestParam double minLon,
+                                                                 @RequestParam double maxLon) {
         return ResponseEntity.ok(reportService.getMarkersInBounds(minLat, maxLat, minLon, maxLon));
     }
 }
