+++ backend/byeoltago-api/src/main/java/com/kakaoimpact/byeoltago_api/repository/ReportClusterRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface ReportClusterRepository extends JpaRepository<ReportCluster, Long> {
     List<ReportCluster> findByTypeId(Integer typeId);  // 신고 유형 필터링
     List<ReportCluster> findByStatusId(Integer statusId);  // 처리 상태 필터링
-    List<ReportCluster> findByLatitudeBetweenAndLongitudeBetween(Float minLat, Float maxLat, Float minLon, Float maxLon);  // 위경도 범위 검색
+    List<ReportCluster> findByLatitudeBetweenAndLongitudeBetween(Double minLat, Double maxLat, Double minLon, Double maxLon);  // 위경도 범위 검색
 }
